Refactor memory and interrupts for bits32/bits64
