+++ src/arch/x86/interrupts/exception/mod.rs
@@ -2,14 +2,14 @@ use core::fmt;
 use x86::shared::irq::{InterruptDescription, EXCEPTIONS};
 
 #[cfg(target_arch = "x86")]
-mod x86;
+mod bits32;
 #[cfg(target_arch = "x86_64")]
-mod x86_64;
+mod bits64;
 
 #[cfg(target_arch = "x86")]
-pub use self::x86::*;
+pub use self::bits32::*;
 #[cfg(target_arch = "x86_64")]
-pub use self::x86_64::*;
+pub use self::bits64::*;
 
 macro_rules! exception {
     ($e:ident, $desc:expr, $func:block) => {
